requirements.txt
Django==1.10.1
argparse==1.2.1
django-crispy-forms==1.6.0
requests==2.11.1
wsgiref==0.1.2
